build: move build process and excalidraw-app dependencies in its own package.json
